refactor: remove link builder
